// frontend/src/types/material.types.ts
export interface Material {
  Id?: number;
  id?: number;
  Name?: string;
  name?: string;
  UnitOfMeasure?: string;
  unitOfMeasure?: string;
  IsActive?: boolean;
  isActive?: boolean;
  IsRequestable?: boolean;
  isRequestable?: boolean;
}

export interface MaterialCategory {
  Id?: number;
  id?: number;
  Name?: string;
  name?: string;
}